# 60-upload-file/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Add Route service to add, import, find and update ticketing route",
        "title": "60-upload-file",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/route": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route"
                ],
                "summary": "Find all route, or find route with filter by from location or to location.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query to search filter by 'from' location",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query to search filter by 'to' location",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AllRoutes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Unauthorized"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMongoDBGet"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route"
                ],
                "summary": "Create new route, this feature need Role Admin.",
                "parameters": [
                    {
                        "description": "request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorCreateRouteValidation"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Unauthorized"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMongoDBUpset"
                        }
                    }
                }
            }
        },
        "/api/v1/route/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route"
                ],
                "summary": "Import Ticketing Route CSV file to App server, this feature only accessible by role admin.",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV route file to import",
                        "name": "routeFile",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RouteImportResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorCreateRouteValidation"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Unauthorized"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMongoDBUpset"
                        }
                    }
                }
            }
        },
        "/api/v1/route/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route"
                ],
                "summary": "Get route by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route id to search",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Unauthorized"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMongoDBGet"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route"
                ],
                "summary": "Update route by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route id to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RouteEditResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorUpdateRouteValidation"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Unauthorized"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorMongoDBGet"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AllRoutes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Route"
                    }
                }
            }
        },
...
        "models.CreateRouteRequest": {
            "type": "object",
            "required": [
                "departureTime",
                "from",
                "price",
                "to"
            ],
            "properties": {
                "departureTime": {
                    "type": "string",
                    "example": "09:00:00"
                },
                "from": {
                    "type": "string",
                    "example": "Malang"
                },
                "price": {
                    "type": "number",
                    "example": 200000
                },
                "to": {
                    "type": "string",
                    "example": "Jakarta"
                }
            }
        },
...
        "models.ErrorCreateRouteValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Field from, to, price, departureTime cannot be empty!"
                }
            }
        },
...
        "models.ErrorMongoDBGet": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "failed to get data from mongodb, {error}"
                }
            }
        },
        "models.ErrorMongoDBUpset": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "failed to insert or update data to mongodb, {error}"
                }
            }
        },
        "models.ErrorUpdateRouteValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Field price, departureTime cannot be empty!"
                }
            }
        },
...
        "models.Route": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6336185fc31ad7ad4022ab87"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2022-09-30T05:12:47.469Z"
                },
                "departureTime": {
                    "type": "string",
                    "example": "09:00:00"
                },
                "from": {
                    "type": "string",
                    "example": "Malang"
                },
                "price": {
                    "type": "number",
                    "example": 200000
                },
                "to": {
                    "type": "string",
                    "example": "Jakarta"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2022-09-30T05:12:47.469Z"
                }
            }
        },
        "models.RouteEditResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Route was updated successfully!"
                }
            }
        },
        "models.RouteImportResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Route data has successfully imported!"
                }
            }
        },
        "models.Unauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "models.UpdateRouteRequest": {
            "type": "object",
            "required": [
                "departureTime",
                "price"
            ],
            "properties": {
                "departureTime": {
                    "type": "string",
                    "example": "09:00:00"
                },
                "price": {
                    "type": "number",
                    "example": 200000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}